従業員管理（EmployeesManagement）、パスワード登録/更新（Password）、タイムレコーダー（TimeRecorder）のアーキテクチャ構造を変更（コントローラクラスに偏ってた業務処理をサービスクラスに移植）。例外処理を追加。また、勤怠履歴（WorkHistory）で、欠勤したシフトの「修正」を押すとNullPointerExceptionが出るエラーの修正。

+++ src/main/java/com/example/webapp/service/impl/EmployeesManagementServiceImpl.java
@@ -2,10 +2,18 @@ package com.example.webapp.service.impl;
 
 import java.util.List;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
 
 import com.example.webapp.entity.Employee;
+import com.example.webapp.exception.DuplicateEmployeeException;
+import com.example.webapp.exception.ForeiginKeyViolationException;
+import com.example.webapp.exception.InvalidEmployeeIdException;
+import com.example.webapp.exception.NoDataException;
+import com.example.webapp.form.EmployeeForm;
+import com.example.webapp.helper.EmployeeHelper;
 import com.example.webapp.repository.EmployeesManagementMapper;
 import com.example.webapp.service.EmployeesManagementService;
 
@@ -14,41 +22,79 @@ import lombok.RequiredArgsConstructor;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class EmployeesManagementServiceImpl  implements EmployeesManagementService{
-	
+public class EmployeesManagementServiceImpl implements EmployeesManagementService {
+
 	private final EmployeesManagementMapper mapper;
+
 	@Override
-	public Employee selectEmployeeById(Integer employeeId) {
-		return mapper.selectById(employeeId);
+	public List<Employee> getAllEmployees() throws NoDataException {
+		List<Employee> allEmployees = mapper.selectAll();
+
+		if (CollectionUtils.isEmpty(allEmployees)) {
+			throw new NoDataException("従業員一覧が取得できませんでした。");
+		}
+		return allEmployees;
+	}
+	
+	@Override
+	public Employee getEmployee(Integer employeeId) throws InvalidEmployeeIdException {
+		Employee employee = mapper.selectById(employeeId);
+		if (employee == null) {
+			throw new InvalidEmployeeIdException("そのIDをもつ従業員データは存在しません");
+		}
+		return employee;
 	}
 
 	@Override
-	public List<Employee> selectAllEmployees(){
-		return mapper.selectAll();
+	public EmployeeForm getEmployeeForm(Integer employeeId) throws InvalidEmployeeIdException {
+		Employee target = mapper.selectById(employeeId);
+		if (target == null) {
+			throw new InvalidEmployeeIdException("そのIDをもつ従業員データは存在しません");
+		}
+		EmployeeForm form = EmployeeHelper.convertEmployeeForm(target);
+		return form;
 	}
-	
+
 	@Override
-	public List<Employee> selectAllIdAndName(){
+	public List<Employee> getAllIdAndName() {
 		return mapper.selectAllIdAndName();
 	}
-	
+
 	@Override
-	public Integer selectEmployeeIdByName(String name) {
+	public Integer getEmployeeIdByName(String name) {
 		return mapper.selectIdByName(name);
 	}
-	
+
 	@Override
-	public void insertEmployee(Employee employee) {
-		mapper.insert(employee);
+	public void insertEmployee(EmployeeForm employeeForm) throws DuplicateEmployeeException {
+		Employee employee = EmployeeHelper.convertEmployee(employeeForm);
+		try {
+			mapper.insert(employee);
+		} catch (DataIntegrityViolationException e) {
+			throw new DuplicateEmployeeException("従業員名が重複しています");
+		}
 	}
-	
+
 	@Override
-	public void updateEmployee(Employee employee) {
-		mapper.update(employee);
+	public void updateEmployee(EmployeeForm employeeForm) throws DuplicateEmployeeException {
+		Employee employee = EmployeeHelper.convertEmployee(employeeForm);
+		try {
+			mapper.update(employee);
+		} catch (DataIntegrityViolationException e) {
+			throw new DuplicateEmployeeException("従業員名が重複しています");
+		}
 	}
-	
+
 	@Override
-	public void deleteEmployeeById(Integer employeeId) {
-		mapper.deleteById(employeeId);
+	public void deleteEmployee(Integer employeeId) throws InvalidEmployeeIdException, ForeiginKeyViolationException {
+		Employee target = mapper.selectById(employeeId);
+		if (target == null) {
+			throw new InvalidEmployeeIdException("そのIDをもつ従業員データは存在しません");
+		}
+		try {
+			mapper.deleteById(employeeId);
+		} catch (DataIntegrityViolationException e) {
+			throw new ForeiginKeyViolationException("そのIDをもつ従業員はシフトに登録されているので削除できません");
+		}
 	}
 }
